+++ core/src/nucleus/ribosome.rs
@@ -3,10 +3,9 @@
 extern crate wabt;
 extern crate wasmi;
 
-use self::wasmi::{
-    Error as InterpreterError, Externals, FuncInstance, FuncRef, ImportsBuilder,
-    ModuleImportResolver, ModuleInstance, RuntimeArgs, RuntimeValue, Signature, Trap, ValueType,
-};
+use self::wasmi::{Error as InterpreterError, Externals, FuncInstance, FuncRef, ImportsBuilder,
+                  ModuleImportResolver, ModuleInstance, RuntimeArgs, RuntimeValue, Signature,
+                  Trap, ValueType};
 
 #[derive(Clone)]
 pub struct Runtime {
@@ -74,8 +73,7 @@ pub fn call(
         .expect("Failed to instantiate module")
         .assert_no_start();
 
-    let memory = main
-        .export_by_name("memory")
+    let memory = main.export_by_name("memory")
         .expect("all modules compiled with rustc should have an export named 'memory'; qed")
         .as_memory()
         .expect("in module generated by rustc export named 'memory' should be a memory; qed")
@@ -90,12 +88,11 @@ pub fn call(
         result: String::new(),
     };
 
-    let i32_result_length: i32 = main
-        .invoke_export(
-            format!("{}_dispatch", function_name).as_str(),
-            &[RuntimeValue::I32(0), RuntimeValue::I32(params.len() as i32)],
-            &mut runtime,
-        )?
+    let i32_result_length: i32 = main.invoke_export(
+        format!("{}_dispatch", function_name).as_str(),
+        &[RuntimeValue::I32(0), RuntimeValue::I32(params.len() as i32)],
+        &mut runtime,
+    )?
         .unwrap()
         .try_into()
         .unwrap();
